refactor: move `Group` to `types`, localize `hydrate_from`, remove `Deref` impls
